add analog preamplification stage
the input amplitude from a guitar (or even a headphone jack) is just too low to reliably detect.

i wanted to avoid this for circuit complexity reasons, but it's (unfortunately? fortunately?) necessary

this does have some major benefits though--
first, the instrumentation amplifier naturally converts the differential input to single-ended, allowing us to halve the input buffer size and avoid doing the differential subtraction in software
second, this lets us use a real buffer for the virtual ground, which should be much more stable than filtered pwm
third, this gives us easily adjustable gain, which is useful, and avoids the question of how exactly to tune the gain in software