implement FFT amplitude equalization
this seems like it shouldn't be necessary...it's quite hard to search for, but i can't find anyone else mentioning a non-flat frequency response from an FFT
and it doesn't make sense--it would make extracting amplitudes from an FFT meaningless if there was some falloff at high frequencies
(i would at least expect one of the FFT implementation papers i've seen to mention correcting for this, if it is a thing)

note for posterity: this happens even with a generated input with no window function. so it really seems like a problem specifically with the FFT, and not with anything else
(there is about 10% attenuation at 4kHz vs. 40Hz, from the input filtering, opamp etc., but that's expected and much smaller than what we're seeing from the FFT [up to 5-10x attenuation])