add holdoff
This suppresses output for some fraction of a wavelength after a pulse, which should make low frequencies more noticeable.

It can still be improved--if we suppress a low frequency, we should extend the holdoff period, so e.g. in the following situation

    ###        <- holdoff
|--*-*---*--|
   ^ ^   ^
   | |   1000Hz
   | 60Hz
   1000Hz

...the current implementation would skip the 60Hz pulse due to the first 1000Hz pulse, and then play the second 1000Hz pulse.

But it shouldn't. The holdoff from the suppressed 60Hz should increase the original holdoff from the first 1000Hz pulse, causing the second 1000Hz pulse to be skipped as well.